% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_mp.R
\name{set_progress_mp}
\alias{set_progress_mp}
\title{multi process progress indicator}
\usage{
set_progress_mp(write_location = NULL)
}
\arguments{
\item{write_location}{where to save progress to}
}
\value{
ProgressMP
}
\description{
Sets up a progress object that writes to a shared file to indicate the total
progress. Progress can be monitored by \code{watch_progress_mp}.
}
\seealso{
watch_progress_mp
}
