% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_progress.R
\name{update_progress}
\alias{update_progress}
\title{updating progress bars}
\usage{
update_progress(.pb = NULL)
}
\arguments{
\item{.pb}{the progress bar object}
}
\value{
the progress bar
}
\description{
Takes care of updating a progress bar and stopping when appropriate
}
