% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_mp.R
\name{watch_progress_mp}
\alias{watch_progress_mp}
\title{watch progress from multi process}
\usage{
watch_progress_mp(n, min_time = 0, watch_location = NULL,
  progress_location = make_kpb_output_decisions())
}
\arguments{
\item{n}{number of times process is running}

\item{min_time}{how long to wait}

\item{watch_location}{where is the progress being written to}

\item{progress_location}{where to write the progress output}
}
\value{
ProgressMPWatcher
}
\description{
sets up a "watcher" function that will report on the progress
of a multi-process process that is being indicated by \code{set_progress_mp}.
}
\seealso{
set_progress_mp
}
